Revert "Fix auction"

+++ Backend/src/main/java/fpt/edu/vn/Backend/DTO/response/BidResponse.java
@@ -5,15 +5,14 @@ import fpt.edu.vn.Backend.pojo.Bid;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 @Data
 public class BidResponse implements Serializable {
     private int bidId;
     private AccountDTO account;
-    private BigDecimal price;
+    private double price;
 
-    public BidResponse(int bidId, AccountDTO account, BigDecimal price) {
+    public BidResponse(int bidId, AccountDTO account, double price) {
         this.bidId = bidId;
         this.account = account;
         this.price = price;
